Move mock data generation code for unit tests into separate generator class to minimize code duplication
