+++ tests/external/generate_attempt_report_test.php
@@ -32,28 +32,13 @@ use quiz_archiver\Report;
 final class generate_attempt_report_test extends \advanced_testcase {
 
     /**
-     * Generates a mock quiz to use in the tests
+     * Returns the data generator for the quiz_archiver plugin
      *
-     * @return \stdClass Created mock objects
+     * @return \quiz_archiver_generator The data generator for the quiz_archiver plugin
      */
-    protected function generate_mock_quiz(): \stdClass {
-        // Create course, course module and quiz.
-        $this->resetAfterTest();
-
-        // Prepare user and course.
-        $user = $this->getDataGenerator()->create_user();
-        $course = $this->getDataGenerator()->create_course();
-        $quiz = $this->getDataGenerator()->create_module('quiz', [
-            'course' => $course->id,
-            'grade' => 100.0,
-            'sumgrades' => 100,
-        ]);
-
-        return (object)[
-            'user' => $user,
-            'course' => $course,
-            'quiz' => $quiz,
-        ];
+    // @codingStandardsIgnoreLine
+    public static function getDataGenerator(): \quiz_archiver_generator {
+        return parent::getDataGenerator()->get_plugin_generator('quiz_archiver');
     }
 
     /**
@@ -123,7 +108,8 @@ final class generate_attempt_report_test extends \advanced_testcase {
         $this->expectException(\required_capability_exception::class);
         $this->expectExceptionMessageMatches('/.*mod\/quiz_archiver:use_webservice.*/');
 
-        $mocks = $this->generate_mock_quiz();
+        $this->resetAfterTest();
+        $mocks = $this->getDataGenerator()->create_mock_quiz();
         $r = $this->generate_valid_request($mocks->course->id, $mocks->quiz->cmid, $mocks->quiz->id, 1);
         generate_attempt_report::execute(
             $r['courseid'],
@@ -165,6 +151,8 @@ final class generate_attempt_report_test extends \advanced_testcase {
         bool $attachments,
         bool $shouldfail
     ): void {
+        $this->resetAfterTest();
+
         if ($shouldfail) {
             $this->expectException(\invalid_parameter_exception::class);
         }
@@ -184,7 +172,7 @@ final class generate_attempt_report_test extends \advanced_testcase {
      */
     public static function parameter_data_provider(): array {
         $self = new self();
-        $mocks = $self->generate_mock_quiz();
+        $mocks = $self->getDataGenerator()->create_mock_quiz();
         $base = $self->generate_valid_request($mocks->course->id, $mocks->quiz->cmid, $mocks->quiz->id, 1);
         return [
             'Valid' => array_merge($base, [
